[AGENT-3621] Adds support for extending DRUM server mode (#698)
* Add new hook file for flask extension

If this file is present in the custom model dir then we will use it to
extend our Flask application

* Add new example for extending Flask server

I took the sklearn example and just added the flask customizer code to it.
I felt like sklearn is pretty simple so that should not confuse from
main part of the example which is the `custom_flask.py` code. I'm using
3rd party Flask extension to help with some of the authentication code to
show that this solution is highly customizable.

* Bump version 1.9.11

* Fix lint

* Add new line

* Change print to logging

* Improve README formatting

* Make it a dev release

* Small renaming tweaks to flask hook code

* Improve comments for sample custom flask hook code

* Add test case for custom flask hook

* Fix changelog

* Fix black

* Bump to 1.9.12

* Fix comment

* Improve readme for new sample

* Rename and update docs

I don't want people confusing the flask extension with a real model template.

* Don't use assert for user driven logic