Update RuntimeParameters.get() to support default val
Similar to the dict.get() method, I think it can be useful to have a
mode that doesn't raise but instead allows a user to specify a default.
The usecase I imagine is with credentials that may not always be required
so a `RuntimeParameters.get('API_KEY', None)` call could be used.