[AGENT-4009] Fix some bugs found during testing RuntimeParams (#732)
* Use `strictyaml` as that is used in other places in the code

It looks like PyYaml isn't a direct dependency but strictyaml is so we should use that.

* Fix setup.py to not need 3rd party deps

* Black-ify

* Allow runtime params in metadata file

We don't need to define the schema in DRUM since the MLOps platform will
pass all runtime params via env vars.

* Bump version

* Fix key name

* Proper use of strictyaml

* Make setup.py more idiomatic

* Revert strictyaml usage

strictyaml doesn't deal with `null` values very simply (requires a schema).
This seems like overkill for code that is only run in local dev. In addition,
adding the PyYAML dependency is currently a noop because it is already
getting pulled in transiently.

* fix yaml.load usage

* Remove unused import