[AGENT-6667]  No monitoring if training data is not available
If the training data is not available for the deployment, drift tracking is
disabled and no point reporting monitoring data.

It is possible that training data is available but drift is disabled, but it
is not possible to check drift for every request.  DRUM will receive an error
code in that case.

Ensured existing tests are running successfully

Added test cases for chat interface

Added test cases for score interface too

Query the champion model package to see if training data is available