[AGENT-6675] Add support for CUSTOM_MODEL_WORKERS param in GPU predictors (#1232)
* Factored health check endpoint into base class

NIM and vLLM basically had the same code so factor out into base class.

* Prepare for new DRUM release

* Prepare for new vLLM release

* Have tests use multiple workers

* Don't use thread to check if alive

The thread is not copied to forks so use the pid of the subprocess.

* Update watchdog code to be multi-worker safe

By moving creating of the watchdog thread to the pre-fork code we can be sure to only have one copy. However it means we need to use files to for IPC.

* Bump workers in tests to more realistic number

* Update Changelog

* Bump NIM images to use new DRUM

* Fix formatting