[AGENT-6734] Revert the check of training data available
The predictionInputs/fromJSON/ needs to accept monitoring data if either the
feature drift is enabled or predictions data collections is enabled.  For Textgen
models, although we don’t have training data and no feature / target drift,
typically we enable the predictions data collections.  We could check in DRUM
if the predictions data collection flag is set on the deployment or not before
reporting monitoring data, but that check has to be made for each report request.
That is not scalable. Hence, we need to let the decision of whether to accept
monitoring data for processing or not to be taken at API controller level and
remove the check of training data from DRUM.

In the context of IR-747, we do want this data still to be reported for monitoring
but as long as we have an HPA this should not be an issue (even if it is failing
with 422)

Debugged the API Controller code to ensure that if we revert this change