[AGENT-6541 AGENT-6865] Allow user to override parts of nim predictor logic (#1294)
* [AGENT-6541] Allow user to override served model name

For LoRA models and multi-container models we need to be more flexible with our served model name. This allows users to set it via runtime param and also we defer to chat completion to set it and only inject the default to maintain backward compatibility.

* Add support for predict (structured) in NIM predictor

* Apply isort to GPU code

* Reorder frameworks

* Lower timeout to match what I see on a g6e.xlarge machine

* Add test coverage for model-name override

* Fixup param names

* Update output assertion for vLLM

* Remove sidecar param

I think we just need dedicated test classes for sidecars because of the extra setup

* Change the way we test nim sidecar

Make it so pytest launches the sidecar so we have more flexibility in testing.

* Handle proxy routes correctly

* Fix fixtures

* Fix direct-access tests for NIM

* Upgrade packages for NIM sidecar

* Add basic test for NemoGuard running as sidecar

* Update changelog

* Prepare new release

* Prepare sidecar image to take new DRUM version