[RAPTOR-12446] Fix invalid test -nz
The test builtin can't take multiple options in a single call, so this yields a
syntax error. Additionally, `-n` (is set) and `-z` (is not set) presents a fun
logic issue. :)

The intent appears to be to check if the related variable is set before
proceeding, so this removes the `z` from `-nz` to accomplish that.