[RAPTOR-14202] Fix annoylib compilation (#1585)
* [RAPTOR-14202] Fix annoylib compilation

nemoguardrails use annoy library which is mostly in C++ with python bindings.
On EKS clusters, running nemoguard in playground and prod results in
"Invalid opcode (core dumped)". This is because while installation of the annoy
library, it is compiled using gcc and the compiled .so has some optimized
instructions which become invalid on EKS - leading to core dumped

With exporting compilation flags specifically for annoy lib, we make sure that
the binary .so is compatible with x86_64 and works on all regression environments

Tested this change for EKS cluster to ensure that nemo guard works and on AKS
cluster to ensure that it is not introducing regression

* Reconcile dependencies, updated IDs, tags

* Fix

---------

Co-authored-by: svc-harness-git2 <svc-harness-git2@datarobot.com>