[BUZZOK-27808] PDIE Genai agents: Use system pip (#1696)
This drops the uv pip install, preferring to rely on the system installed pip
and uv.  This is currently to address CVE-2025-8869, which upstream pip has
fixed but has not released yet, and its release date is undetermined at this
time.

Chainguard however has fixed this in their images, so pip installing from pypi
is re-introducing the vulnerability into our images. Relying on system pip
allows us to use the fixes from Chainguard.