[RAPTOR-14743] pdie rlang: rely on system pip
This is to address CVE-2025-8869. Chainguard has fixed this in the base image
but the fix hasn't been released in upstream pip yet, so by re-installing pip
from upsteam, we are reintroducing the vulnerability that chainguard has fixed.

This just relies on the syste pip for the virtualenv setup so the cve stays
patched.