[RAPTOR-14742] PDIE: All without pip (#1703)
* [RAPTOR-14742] PDIE: All without pip

This is for all relevant public dropin environments. This drops the pip install
and update that happens, preferring to rely on the system installed pip. This
is currently to address CVE-2025-8869, which upstream pip has fixed but has not
released yet, and its release date is undetermined at this time.

Chainguard however has fixed this in their images, so pip installing from pypi
is re-introducing the vulnerability into our images. Relying on system site
packages, especially for pip, allows us to use the fixes from Chainguard
without re-introducing them.

* Reconcile dependencies, updated IDs, tags

---------

Co-authored-by: svc-harness-git2 <svc-harness-git2@datarobot.com>