[RAPTOR-14944] Fix libxml in r_lang for chainguard (#1723)
Chainguard's upstream libxml2 package seems to symlink wrong, breaking the
installation of r-lang xml2, which prevents the installation of tidyverse and
devtools.

While we wait on Chainguard to fix this upstream, this implements a temporary
workaround which deletes the broken symlink, and relinks it to the existing
correct symlink for libxml2.so.16.

This is safe to keep once chainguard fixes it upstream, as it only executes
when the broken symlink is detected, but this should be removed once Chainguard
fixes things.