[RAPTOR-14966] Test Functional by Framework change build job (#1731)
This changes the `test_functional_by_framework_multisteps` pipeline to use the
new `Build_and_Publish_ExecEnv` pipeline. This pipeline uses the newly
established standard build process for exec envs, which will ensure that images
are properly labelled and built with the same process regardless of which
project builds the EE.

Also note that this stops building and testing the `.local` images as those are
not part of the production release process, so we don't need to build and test
them with every change.