[RAPTOR-15085] PDIE r_lang: Fix libnghttp copy
Some tests started to fail with the absence of the `pack` package for r lang.
This package is installed during testing, but to do so requires the libnghttp
library. However, this was recently upgraded in the underlying chainguard image
from libnghttp2 to libnghttp3, so this changes the COPY directive to grab
version 3 so pack installs successfully.