Fix bug from the reconcile stage
It seems when the bson string contains only numbers, the python code in the
reconcile stage reads it as an int, and in the odd case that it ends with a
`.*e[0-9]+`, it treats it like a scientific notation.

Soooo, this just regenerates that env id for sklearn. Will fix the reconcile
stage later.