[-] Autobump env_info version IDs (#1770)
* [-] Autobump env_info version IDs

This is an execution environment version ID auto bump to trigger automated
image rebuilds for rebase on the latest base image builds.

[Execution:Custom_Models/execution_environment_installer/EE_Chart_Bump_All_Environment_Versions/nWPt75XoS5-nkqqau7YkHw](https://app.harness.io/ng/account/oP3BKzKwSDe_4hCFYw_UWA/module/ci/orgs/Custom_Models/projects/execution_environment_installer/pipelines/EE_Chart_Bump_All_Environment_Versions/executions/nWPt75XoS5-nkqqau7YkHw)

* Reconcile dependencies, updated IDs, tags

* Fix bug from the reconcile stage

It seems when the bson string contains only numbers, the python code in the
reconcile stage reads it as an int, and in the odd case that it ends with a
`.*e[0-9]+`, it treats it like a scientific notation.

Soooo, this just regenerates that env id for sklearn. Will fix the reconcile
stage later.

---------

Co-authored-by: GenAI Git Bot <genai-systems-devs@datarobot.com>
Co-authored-by: Aaron Ball <aaron.ball@datarobot.com>