[RAPTOR-15278] Bump keras to fix a few CVEs (#1814)
* [RAPTOR-15278] Bump keras to fix a few CVEs

Note the additional pin of `scikit-learn >=1.6.1,<1.7`. This is because in
1.6.0 scikit-learn modified its APIs where tagging is concerned. Fortunately in
the 1.6.x series, these modifications just produce warnings, but in 1.7 and up
they produce errors, which is what was causing all the test failures during
past attempted upgrades.

As this is just maintenance of an already cut release, we will focus on
bringing master fully up-to-date, but will leave this branch maintained.

* Reconcile dependencies, updated IDs, tags

---------

Co-authored-by: GenAI Git Bot <genai-systems-devs@datarobot.com>