[RAPTOR-15281] Bump sklearn to fix a few CVEs (#1815)
* [RAPTOR-15281] Bump sklearn to fix a few CVEs

Note the additional pin of `scikit-learn >=1.6.1,<1.7`. This is because in
1.6.0 scikit-learn modified its APIs where tagging is concerned. Fortunately in
the 1.6.x series, these modifications just produce warnings, but in 1.7 and up
they produce errors, which is what was causing all the test failures during
past attempted upgrades.

This also upgrades requirements_test.txt scikit-learn to match the above pin to
avoid installing two versions at the same time, as well as unpinning scipy for
the same reason. This resolves a lot of testing issues during upgrades.

* Add our own stderr filter for FutureWarning

* Reconcile dependencies, updated IDs, tags

---------

Co-authored-by: GenAI Git Bot <genai-systems-devs@datarobot.com>