[RAPTOR-15457] PDIE: Regen all requirements.txt (#1822)
* [RAPTOR-15457] PDIE: Regen all requirements.txt

This is to pull in the latest datarobot-drum to resolve a CVE with apache log4j
core. This should also address the remaining werkzeug and urllib CVEs.

Note the additional pin of `scikit-learn >=1.6.1,<1.7`. This is because in
1.6.0 scikit-learn modified its APIs where tagging is concerned. Fortunately in
the 1.6.x series, these modifications just produce warnings, but in 1.7 and up
they produce errors, which is what was causing all the test failures during
past attempted upgrades.

This also upgrades requirements_test_unit.txt scikit-learn to match the above
pin to avoid installing two versions at the same time. This resolves a lot of
testing issues during upgrades.

* Reconcile dependencies, updated IDs, tags

---------

Co-authored-by: GenAI Git Bot <genai-systems-devs@datarobot.com>