[RAPTOR-15282] Bump PDIE xgboost for urllib
This should address a couple of CVEs for this image.

Note the additional pin of `scikit-learn >=1.6.1,<1.7`. This is because in
1.6.0 scikit-learn modified its APIs where tagging is concerned. Fortunately in
the 1.6.x series, these modifications just produce warnings, but in 1.7 and up
they produce errors, which is what was causing all the test failures during
past attempted upgrades.

Also note the pin added to xgboost, keeping it below the 3.1.x series. This is
because they
[removed](https://github.com/dmlc/xgboost/blob/master/doc/changes/v3.1.0.rst#removing-deprecated-features)
the binary serialization format from xgboost in that version. Since this is
just a maintenance release now, we'll continue on the older version.